requirements.txt
aerich==0.7.1
aiohttp==3.8.4
pytz==2023.3
Requests==2.31.0
sanic==23.3.0
sanic_ext==23.3.0
tortoise==0.1.1
python-dotenv==1.0.0
cryptography==40.0.1
urllib3==1.26.14
sanic_jwt==1.8.0